refactor: move summarization chain to chains package
